+++ react/CategoryCard.js
@@ -1,0 +1,3 @@
+import CategoryCard from './components/CategoryCard/index'
+
+export default CategoryCard